refactor/add tests
